requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
openai
requests